+++ software/C/PV_DC_driver/src/MEASUREMENTS/measurements.h
@@ -3,7 +3,7 @@
 
 #include <inttypes.h>
 
-void measurements_init(uint16_t voltage_v_ref, uint16_t voltage_multiplier, uint16_t current_v_ref);
+void measurements_init(uint16_t voltage_v_ref, float voltage_multiplier, uint16_t current_v_ref);
 double get_voltage_value(int voltage, float duty_value, float pwm_duty_resolution);
 double get_current_value(int voltage, float duty_value, float pwm_duty_resolution);
